refactor : Constants와 CarFixtures의 패키지 정리
