+++ src/main/java/racingcar/constants/Constants.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.constants;
 
 public class Constants {
