+++ src/test/java/racingcar/model/CarTest.java
@@ -2,7 +2,7 @@ package racingcar.model;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
 import static org.assertj.core.api.Assertions.assertThat;
-import static racingcar.CarFixtures.POBY;
+import static racingcar.fixtures.CarFixtures.POBY;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
